FIX: Only allow PATCH requests when accessing `…/use`
The PATCH method is used to update the text bodies of existing canned replies (via `update`), and also to keep a count of the number of times a canned reply is used (via `use`).

Access control to the various endpoints is controlled via the `CannedRepliesConstraint`, and it was allowing any GET or PATCH requests through if `can_use_canned_replies?` was true. This is why it would allow normal users to edit existing canned replies, but not delete them, or create new replies.

This commit adds an additional condition, to check that the request will end up on the `/use` endpoint. Additionally, tests have been added to ensure the correct logic is used when a normal user tries to edit posts.