FIX: @computed in CannedTagDrop was causing test to fail
The @computed used in the component CannedTagDrop was causing the test "Filtering by Tags"
to fail. It seems that the property topTags wasn't being recomputed when then property
availableTags was updated.

It's odd because this behavior is manifesting only when running the tests. The component remained
functional in the browser. Replacing the @computed decorator with @discourseComputed and ajusting
the code solved the problem.